# Social Media Platform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media Platform API
  version: 1.0.0
  description: REST API documentation for the Selamat social media platform.
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.selamat.example.com/api
    description: Production server placeholder
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get the current authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/search:
    get:
      tags: [Users]
      summary: Search active users
      operationId: searchUsers
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: Users returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get a user by ID
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /posts/{postId}/reactions:
    put:
      tags: [Reactions]
      summary: Add or replace the current user's reaction to a post
      operationId: reactToPost
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionRequest'
      responses:
        '200':
          description: Reaction saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Reactions]
      summary: Remove the current user's reaction from a post
      operationId: removePostReaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204':
          description: Reaction removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Authentication is required or the token is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    User:
      type: object
      required: [id, username, email, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        status:
          $ref: '#/components/schemas/UserStatus'
        roles:
          type: array
          items:
            type: string
          example: [USER]
        createdAt:
          type: string
          format: date-time
...
    ReactionRequest:
      type: object
      required: [reactionType]
      properties:
        reactionType:
          $ref: '#/components/schemas/ReactionType'
    Reaction:
      type: object
      required: [id, userId, reactionType, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
          nullable: true
        commentId:
          type: string
          format: uuid
          nullable: true
        reactionType:
          $ref: '#/components/schemas/ReactionType'
        createdAt:
          type: string
          format: date-time
...
    UserSummary:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed.
        details:
          type: array
          items:
            type: string
    UserStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, DELETED]
...
    ReactionType:
      type: string
      enum: [LIKE, LOVE, LAUGH, SAD, ANGRY]
...